requirements.txt
torch
torchvision
numpy
matplotlib
streamlit
Pillow
scikit-learn


